+++ android/app/src/main/java/com/awesomeapp/OndataPass.kt
@@ -1,5 +1,0 @@
-package com.awesomeapp
-
-interface OndataPass {
-    fun onDataPass(data: String?)
-}