+++ src/components/navigation/AppNavigation.tsx
@@ -2,11 +2,11 @@ import { createDrawerNavigator } from '@react-navigation/drawer';
 import { createStackNavigator } from '@react-navigation/stack';
 import { Dimensions } from 'react-native';
 import * as PropTypes from 'prop-types';
-import AboutUsScreen from '../../screens/staticScreens/AboutUsScreen';
-import SplashScreen from '../../screens/staticScreens/SplashScreen';
-import MainScreen from '../../screens/MainScreen';
-import CustomDrawerNavigator from './CustomDrawerNavigator';
-import LoginScreen from '../../screens/auth/LoginScreen';
+import AboutUsScreen from 'src/screens/staticScreens/AboutUsScreen';
+import SplashScreen from 'src/screens/staticScreens/SplashScreen';
+import MainScreen from 'src/screens/MainScreen';
+import CustomDrawerNavigator from 'src/components/navigation/CustomDrawerNavigator';
+import LoginScreen from 'src/screens/auth/LoginScreen';
 
 const { width, height } = Dimensions.get('screen');