+++ src/screens/staticScreens/SplashScreen.tsx
@@ -1,9 +1,9 @@
 import { Text } from 'react-native';
-import I18n from '../../../shared/I18n/I18n';
-import Header from '../../components/sections/header';
-import HamburgerButton from '../../../shared/components/buttons/HamburgerButton';
-import { Images } from '../../../shared/themes';
-import { CustomScroll, LogoImage, SafeView } from '../../../shared/styled/global.ltr.styles';
+import I18n from 'shared/I18n/I18n';
+import Header from 'src/components/sections/header';
+import HamburgerButton from 'shared/components/buttons/HamburgerButton';
+import { Images } from 'shared/themes';
+import { CustomScroll, LogoImage, SafeView } from 'shared/styled/global.ltr.styles';
 
 const SplashScreen = (): JSX.Element => {
   return (